Create admin database first to streamline install
The creation of the admin database asks for a password from the user
and when that step appears last in the create-db.sh process, the user
must wait for the other database processes to finish before
interacting with the script. Moving the admin db creation to the
beginning of the process makes it so the user can enter the password
at the beginning and then they don't have to interact with it anymore.