Use packaging.version for version parsing
There was an import problem with the previous pkg_resources-based
solution for comparing versions, so this commit moves to use the
'packaging' module for comparing versions. I think the end result is a
bit easier to understand and it is better documented, although it does
incur a new dependency (but to be fair, pkg_resources is also a
dependency, it's just included with setuptools I think).