Add gwadoc info to relations on hover & click
The short description and short example are shown in the following
format: `{definition} (E.g. {example})` if present. If neither are
defined in `gwadoc`, then `(no description; click for more info)` is
shown. Clicking on a relation name goes to the gwadoc hosted
documentation, as before, but now the target is named instead of
`_blank` so successive clicks update the target window. Note that the
hover and click behavior only happens on the relations list of concept
pages (`omw_concept.html`).

Resolves #54