Add statistics for sense freq, improved core docs
The wordnet stats page now shows the percentage of senses that have a
non-zero frequency and the total number of instances.

The wordnet stats page now links to the core documention, and the core
documentation allows you to download a list of core ilis.

Also validate-wn now properly imports updateLabels and the README
tells you how to run validate-wn.