Make get-examples script output more info
It now includes the CILI ids and synset ids.

Also the script now can take a -n N argument to change how many
results per relation to return, and the lg argument is actually used
in the query to select a language.

Also the TSV header line is printed first.